Change structure; kept backwards compatibility
