+++ src/examples/hexapod.cpp
@@ -4,7 +4,7 @@
 #include <robot_dart/control/hexa_control.hpp>
 
 #ifdef GRAPHIC
-#include <robot_dart/graphics/graphics.hpp>
+#include <robot_dart/gui/osg/graphics.hpp>
 #endif
 
 #include <dart/collision/bullet/BulletCollisionDetector.hpp>
@@ -33,8 +33,8 @@ int main()
 
     robot_dart::RobotDARTSimu simu(0.001);
 #ifdef GRAPHIC
-    simu.set_graphics(std::make_shared<robot_dart::graphics::Graphics>(simu.world()));
-    std::static_pointer_cast<robot_dart::graphics::Graphics>(simu.graphics())->look_at({0.5, 3., 0.75}, {0.5, 0., 0.2});
+    simu.set_graphics(std::make_shared<robot_dart::gui::osg::Graphics>(simu.world()));
+    std::static_pointer_cast<robot_dart::gui::osg::Graphics>(simu.graphics())->look_at({0.5, 3., 0.75}, {0.5, 0., 0.2});
 #endif
     simu.world()->getConstraintSolver()->setCollisionDetector(dart::collision::BulletCollisionDetector::create());
     simu.add_floor();
